Move ClaimMetrics to ReviewStats
